Added temporary boost workaround - some assimp features work with reduced functionality in this case. Added AC-loader, WIP version. PLY loader is now able to load models from blender, test model added. Refactoring. Added FindInvalidData step. Added support for precompiled headers, the release builds in VC8 are configued to use PCH now. Added separate makefile for mingw, no -FPic warning anymore, -clear works now.
git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@176 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ code/MDLLoader.cpp
@@ -42,23 +42,12 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 /** @file Implementation of the main parts of the MDL importer class */
 
 // internal headers
-#include "MaterialSystem.h"
+#include "AssimpPCH.h"
+
 #include "MDLLoader.h"
 #include "MDLDefaultColorMap.h"
 #include "MD2FileData.h" 
-#include "qnan.h" 
-#include "ByteSwap.h" 
 
-// public ASSIMP headers
-#include "../include/DefaultLogger.h"
-#include "../include/aiScene.h"
-#include "../include/aiAssert.h"
-#include "../include/IOStream.h"
-#include "../include/IOSystem.h"
-#include "../include/assimp.hpp"
-
-// boost headers
-#include <boost/scoped_ptr.hpp>
 
 using namespace Assimp;
 
@@ -1712,7 +1701,7 @@ void MDLImporter::AddBonesToNodeGraph_3DGS_MDL7(const MDL::IntBone_MDL7** apcBon
 		aiNode* pcNode = pcParent->mChildren[qq++] = new aiNode();
 		pcNode->mName = aiString( pcBone->mName );
 
-		this->AddBonesToNodeGraph_3DGS_MDL7(apcBones,pcNode,i);
+		this->AddBonesToNodeGraph_3DGS_MDL7(apcBones,pcNode,(uint16_t)i);
 	}
 }
 // ------------------------------------------------------------------------------------------------
